Move to new NodeCG install method
