+++ .gitignore
@@ -28,3 +28,8 @@ extension/*
 dashboard/*
 graphics/*
 vite.config.mts.*
+
+# NodeCG runtime files
+db/*
+cfg/*
+assets/*
