+++ src/browser/dashboard/example.tsx
@@ -1,6 +1,6 @@
 import Button from '@mui/material/Button';
 import exampleImage from './image.png';
-import { useReplicant } from 'use-nodecg';
+import { useReplicant } from '@nodecg/react-hooks';
 import { ExampleReplicant } from '../../types/generated';
 import { DashboardThemeProvider } from './components/DashboardThemeProvider';
 import { createRoot } from 'react-dom/client';
@@ -8,13 +8,12 @@ import { createRoot } from 'react-dom/client';
 const App = () => {
   const [exampleReplicant, setExampleReplicant] = useReplicant<ExampleReplicant>(
     'exampleReplicant',
-    { exampleProperty: '' }
   );
 
   return (
     <DashboardThemeProvider>
       <img src={exampleImage} />
-      <p>Current example replicant value: {exampleReplicant.exampleProperty}</p>
+      <p>Current example replicant value: {exampleReplicant?.exampleProperty}</p>
       <Button
         variant="contained"
         onClick={() => {
