+++ src/extension/index.ts
@@ -1,6 +1,6 @@
 import { Configschema } from 'src/types/generated';
 import { set } from './util/nodecg';
-import NodeCG from '@nodecg/types';
+import type NodeCG from 'nodecg/types';
 
 export default (nodecg: NodeCG.ServerAPI<Configschema>) => {
   set(nodecg);
