+++ src/extension/util/nodecg.ts
@@ -1,13 +1,12 @@
-import type NodeCG from '@nodecg/types';
+import type NodeCG from 'nodecg/types';
 import { Configschema } from '../../types/generated';
 
-let nodecg: NodeCG.ServerAPI<Configschema>;
-export type NodeCGServer = NodeCG.ServerAPI<Configschema>;
+let nodecg: NodeCG.ServerAPI<Configschema>
 
 export function set(ctx: NodeCG.ServerAPI<Configschema>): void {
-  nodecg = ctx;
+  nodecg = ctx
 }
 
 export function get(): NodeCG.ServerAPI<Configschema> {
-  return nodecg;
+  return nodecg
 }
