+++ src/types/augment-window.d.ts
@@ -3,10 +3,11 @@
  * the NodeCG types, but allows us to automatically receive the configuration types.
  */
 
-import { NodeCGAPIClient } from '@nodecg/types/client/api/api.client';
+import { NodeCGAPIClient } from 'node_modules/nodecg/out/client/api/api.client';
+import type NodeCG from 'nodecg/types';
 import { Configschema } from './generated';
 
 declare global {
   let NodeCG: typeof NodeCGAPIClient;
-  let nodecg: NodeCGAPIClient<Configschema>;
+  let nodecg: NodeCG.ClientAPI<Configschema>;
 }
